+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to the "koby" extension will be documented in this file.
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## [0.1.17]
+- Remove all this occurences according default lint
+- Update Alfreed template
+- Update stateful template
+- Update Hermep template
+- Remove mocktail
+
 ## [0.1.16]
 - Upgrade template to works with latest alfreed version.
 - Support gate for Alfreed page test.